brick_breaker_game/backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
SQLAlchemy==2.0.21
python-dotenv==1.0.0
Werkzeug==2.3.7